state: Move Authorization into its own header
The EIP-7702 authorization tuple has its own signing hash and signature
recovery, which do not belong in transaction.hpp. Give it a home first.

Order the fields as the tuple is specified: chain_id, address, nonce,
y_parity, r, s; and put signer, which is not part of it, at the end.

y_parity becomes uint8_t: EIP-7702 bounds it to one byte, and a value
above 1 invalidates the tuple, not the transaction, so the field must
still be able to hold it. The state test loader now rejects a wider one.
